last updates and bug fixes after using for a while in deployment

+++ flaskr/static/singlerecipe.js
@@ -5,20 +5,24 @@ var modifyRecipeBtn = document.querySelector('.modify-recipe-btn')
 document.addEventListener('click', event =>{
     // delete recipe completely
     if (event.target == deleteRecipeBtn){
-        let recipeIDItem =  deleteRecipeBtn.getAttribute('name')
-        let recipeID = recipeIDItem.slice('recipeID='.length)
-        fetch('/delete-recipe/' + recipeID, 
-            {
-                method: 'DELETE',
-                headers: {
-                    'Accept' : 'application/json',
-                    'Content-Type': 'application/json'
-                    }
-            })
-        window.location.replace("/my-recipes");
+        if (confirm("Ar tikrai norite ištrinti šį receptą?") == true){
+            let recipeIDItem =  deleteRecipeBtn.getAttribute('name')
+            let recipeID = recipeIDItem.slice('recipeID='.length)
+            fetch('/delete-recipe/' + recipeID, 
+                {
+                    method: 'DELETE',
+                    headers: {
+                        'Accept' : 'application/json',
+                        'Content-Type': 'application/json'
+                        }
+                })
+            window.location.replace("/");
+        }
     }
     // modify recipe. Works together with recipes.js functions
     if (event.target == modifyRecipeBtn){
+        
+
         let recipeToModifyName = modifyRecipeBtn.getAttribute('name')
         let recipeToModifyId = recipeToModifyName.slice('recipeID='.length)
         sendGetRequestAwaitforResponse('/create-ingredient-dictionary?recipeID=', recipeToModifyId).then(
@@ -35,14 +39,20 @@ document.addEventListener('click', event =>{
             let prefixlength = 'heartID'.length
             let heartID = heartBtn.id
             let extractedRecipeID=heartID.substring(prefixlength)
-            if (heartBtn.classList.contains('favorited')){
-                heartBtn.classList.remove('favorited')
-                favoritesTableAddOrRemove(extractedRecipeID, 'NO')
-            }
-            else{
-                heartBtn.classList.add('favorited')
-                favoritesTableAddOrRemove(extractedRecipeID, 'YES')
-            }
+            let response = ''
+            heartBtn.classList.contains('favorited') ? response= 'NO' : response = 'YES'
+            favoritesTableAddOrRemove(extractedRecipeID, response)
+            heartBtn.classList.toggle('favorited')
+
+
+            // if (heartBtn.classList.contains('favorited')){
+            //     heartBtn.classList.remove('favorited')
+            //     favoritesTableAddOrRemove(extractedRecipeID, 'NO')
+            // }
+            // else{
+            //     heartBtn.classList.add('favorited')
+            //     favoritesTableAddOrRemove(extractedRecipeID, 'YES')
+            // }
         }
     }
 })
